requirements.txt
Django==1.8.7
django-admin-bootstrapped==2.5.5
django-crispy-forms==1.5.2
django-jet==0.1.4
django-wysiwyg-redactor==0.4.9
Pillow==3.0.0
wheel==0.24.0
